refactor: UrlMapping 엔티티 생성자 수정
