+++ src/main/java/com/tangerine/urlshortener/global/exception/NotFoundMappingException.java
@@ -1,4 +1,4 @@
-package com.tangerine.urlshortener.global;
+package com.tangerine.urlshortener.global.exception;
 
 public class NotFoundMappingException extends RuntimeException {
